REMOVE T-MOBILE THEME ENGINE (3/3)
As it stands right now, this is only functional on 2-3 devices and for some reason, is not on about 4-5 devices.

I've never seen anything like this before but it is obvious that there are some conflicts that need to be fixed.

With that said, I want it out of DU for now since is obviously not stable and until we can fix the conflicts.

Change-Id: If9d09704956405cc59028d22d70c641f2c69a666