Let's adjust this a bit (see note)
Currently the only feature we wanted to hide in the navbar category was the left hand
navbar because it really doesn't apply to tablets but only phones.

With the current implementation, it was removing the category and anything underneath it.

This was not a problem until we wanted to add the button glow feature and so it removed
that as well and even so, on phones when DPI was set too low.

This fixes that.....

Change-Id: I9ea03ca961ad94b14a2537b51784ee96d48b44a2