Add back blacklist settings. [Squashed]
Cm originally had this squashed, but with the mess that is cm-13 even theyre squashed became unsquashed and you had to dig in like 10 commits that had nothing to do with blacklist to find all the code. thats what happens when one person merges fwbase and another setting.. c'mon man organize your shit!

Change-Id: I45d6c90bd584d65c83504a0fdca96f69f4ee7d94