Bring back the 'AOSP look'
Lot of people feel like the switches force you to lose the 'AOSP look'
and in turn are stuck with the 'CM look'. This is due to the dashboard
mobile tile made visible while the preference under 'More' was deleted.

Now you can have both. The 'AOSP look' while switches are off and the
'CM look' while the switches are on.

Change-Id: I51c958b6aa50f108fef27e263ce6b2186bcc3f4d