Organize Ambient display options
- Remove Ambient brightness options from the OP3T (see note)

Note:
While these options work on the OP3T, is not as clean as I would like.

In order for them to work, the user would have to go into our Doze 
package and enable 'Pick up'. The user would then have to exit out 
of that package and then enable these options. Is just sloppy and 
feels out of place.

I'm going to submit another patch in where I add this to the Doze package
to give it a more native feel. Right now I'm not getting that.

Change-Id: Ied940f99f9ee39b01f6164f968cb736d8b78f073