Decouple hardware revision preference in "Model & hardware"
On some devices this field is blank/unknown. Instead of making some shit up,
we're going to hide it on these devices. The problem here was that Google and
infinite wisdom decided to use the same bool for category and all the preferences.

So we had to decouple it so we can just hide the hardware revision preference.

Change-Id: I52f7336b86730fae6a0c3b4ee7d49ac46d1f562e