Ensure that 'Tap to wake' is off on all devices
This change is being made for consistency across all devices.

Some unofficial builders have gone ahead and enabled this via
the overlay but it just causes problems where users think that
we're 'missing' this on our official builds. Instead of aruging
with them and explaining the same thing 400 times a day, we're
just shutting the switch off from our side :-)

Change-Id: Ic30509c3c35663bfc0cd0566430ddfc34551bf75