Do not allow disabling the default keyboard
It's not obvious that disabling it after installing another keyboard is
a very bad idea. Other keyboards rarely support Direct Boot and won't
work for entering the password, forcing recovery by plugging in a
physical keyboard. Even if they do support Direct Boot it's a bad idea
to rely on third party apps for basic functionality to keep working.

**DU Edits**
- Change to GBoard, since that's what we ship with

Change-Id: I35b92efa0b783de79ba232a5e79e50d53070ef61